tiktok-scraper-python/requirements.txt
# TikTok Scraper Python Dependencies
playwright==1.40.0
psycopg2-binary==2.9.9
aiohttp==3.9.1
python-dotenv==1.0.0
asyncio==3.4.3
pandas==2.1.4
numpy==1.26.4
pydantic==2.5.2
loguru==0.7.2
tenacity==8.2.3
aiofiles==23.2.1
python-dateutil==2.8.2 